pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-forecast-london"
version = "0.1.0"
description = "Short-term traffic congestion prediction for London using TomTom, TfL, and DfT APIs."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
